refactor: convert charge mongoose model to a schema
